ClarifEye.AI/TrafficLightsDetector/requirements.txt
tensorflow>=2.11
tensorflow-datasets>=4.9.2
numpy>=1.21
pillow>=9.0.0
fastapi >=0.95.2
uvicorn >=0.22.0
python-multipart>=0.0.5